# feat: kernel pooling to reduce latency