cc @havok2063 @timkpaine 

Inpired by #726, related #30 


This creates a pool of kernel, of size `pool_size.get(kernel_name, pool_size_default)` (default=1) that are always kept in the background to reduce latency, *not* max throughput or server load.

When the kernel is started, a file with `{prefix}/voila_kernel_pool_init_{kernel_name}.{ext}` (e.g. `/Users/maartenbreddels/.jupyter/voila_kernel_pool_init_python3.py`) will be executed in the kernel, and if the kernel language is `python`, also the modules listed in `python_imports` will be imported, the most commonly anticipated usecase. (Note: this all happens before any notebooks are considered).

The effect of executing code beforehand (mainly importing) is to speed up the execution of the notebook. This *does not reduce the load* since the same amount of code will be executed, just distributed differently through time.

To make sure the cwd and env arguments of `start_kernel` are effective, we have some python update code, so this will currently only work with python. (*Note: env and cwd issues will also affect kernel forking, or #726*)

Example notebook:
```python
import jdaviz
```

Before (>3 seconds):
```
$ voila notebooks/preheat.ipynb --no-browser --PoolMappingKernelManager.pool_size_default=0
$ time curl http://localhost:8866/
curl http://localhost:8866/  0.01s user 0.01s system 0% cpu 3.727 total
```

After (<0.7 seconds):
```
$ voila notebooks/preheat.ipynb --no-browser  --PoolMappingKernelManager.python_imports=jdaviz --PoolMappingKernelManager.pool_size_default=5
$ time curl http://localhost:8866/
curl http://localhost:8866/  0.01s user 0.01s system 2% cpu 0.609 total
```

We can think about upstreaming this in jupyter_server.

TLDR: this speeds up the time to show a notebook with voila from >3 seconds to <0.7 seconds by having a warm kernel, with preimported modules.