@havok2063 2021-01-29
@maartenbreddels This looks very promising and pretty straightforward to implement.  Presumably `python_imports` accepts a list of packages?  Can all of these settings be controlled via a Voila config file?