@maartenbreddels 2021-01-29
Yes indeed, in json it should be something like:
```
{
 "PoolMappingKernelManager": {
  "python_imports": ["numpy", "jdaviz"],
  "pool_size_default": 5
 }
}
```