@havok2063 2021-02-05
Ok, node upgraded.  It turns out my conda environment where I run all my voila stuff was quite outdated.  Now I get the following error `AttributeError: module 'nbconvert.exporters.templateexporter' has no attribute 'DEV_MODE'` when I run the following voila command 
```
voila --no-browser --template=embed --enable_nbextensions=True --Voila.tornado_settings="{'allow_origin': '*'}" --port=8000 --MappingKernelManager.cull_interval=60 --MappingKernelManager.cull_idle_timeout=120 --PoolMappingKernelManager.python_imports=jdaviz --PoolMappingKernelManager.pool_size_default=5
```