@maartenbreddels 2021-02-05
That's odd, because that means you use nbconvert 5, while we require
nbconvert 6. If using multiple environments, check against shell PATH
caching, and run `which voila` (rehash or hash -r to fix that, or
reactivate the environment)

(from mobile phone)

On Fri, Feb 5, 2021, 19:09 Brian Cherinka <notifications@github.com> wrote:

> Ok, node upgraded. It turns out my conda environment where I run all my
> voila stuff was quite outdated. Now I get the following error AttributeError:
> module 'nbconvert.exporters.templateexporter' has no attribute 'DEV_MODE'
> when I run the following voila command
>
> voila --no-browser --template=embed --enable_nbextensions=True --Voila.tornado_settings="{'allow_origin': '*'}" --port=8000 --MappingKernelManager.cull_interval=60 --MappingKernelManager.cull_idle_timeout=120 --PoolMappingKernelManager.python_imports=jdaviz --PoolMappingKernelManager.pool_size_default=5
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/voila-dashboards/voila/pull/820#issuecomment-774196976>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AANPEPJOLNCI4OW5FSZD4ZLS5QX45ANCNFSM4WXRUG2Q>
> .
>