@vidartf 2021-02-08
I see we are too much in sync, but uncoordinated 😄 . I started some work here as well: https://github.com/vidartf/hotpot_km