@havok2063 2021-02-08
Yeah I'm using `nbconvert=6.07` in all my environments, so I'm not sure what's going on. Clearing the shell cache didn't help. 
 Let me try creating a brand new environment.