@havok2063 2021-02-11
Ok, I managed to sort my environments and package versions.  I managed to get it working however the existing `jdaviz` has a pinned dependency on `traitlets<5` which is incompatible with this PR.  Once I disabled that dependency version, I was able to run everything fine.  As this PR gets finalized, we'll need to coordinate with the `jdaviz` folks to see why they've pinned `traitlets` and see if we can remove it.  

Regarding the JSON configuration for running the Voila as a standalone server, which config file should contain the settings and where is it located?  The docs only mention a `config.json` for configuring templates.  Is there a top level config.json that should live in `share/jupyter/voila/`? I've found several `voila.json` config files inside `/etc/jupyter_server_config.d`and `/etc/jupyter_notebook_config.d`, but are those for running Voila as a Jupyter server extension?