@maartenbreddels 2021-02-19
For voila standalone, we use the traitlet system to find files, which are named `voila.(json|py)` is several locations, most easily found by passing the `--debug` on the command line, e.g.:
```
[Voila] Looking for voila in /etc/jupyter
[Voila] Looking for voila in /usr/local/etc/jupyter
[Voila] Looking for voila in /Users/maartenbreddels/miniconda3/envs/jdaviz-dev/etc/jupyter
[Voila] Looking for voila in /Users/maartenbreddels/.jupyter
[Voila] Loaded config file: /Users/maartenbreddels/.jupyter/voila.py
[Voila] Looking for voila in /Users/maartenbreddels/src/spacetelescope/jdaviz
[Voila] Loaded config file: /Users/maartenbreddels/src/spacetelescope/jdaviz/voila.json
```