@maartenbreddels 2021-02-23
Replaced by https://github.com/voila-dashboards/voila/pull/841 and  https://github.com/voila-dashboards/hotpot_km/pull/1